# backend_go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/enrollments/{id}": {
            "get": {
                "description": "Retrieve a specific user enrollment by ID",
                "tags": [
                    "enrollments"
                ],
                "summary": "Get user enrollment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enrollment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing user enrollment",
                "tags": [
                    "enrollments"
                ],
                "summary": "Update user enrollment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enrollment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateEnrollmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/enrollments/{id}/cancel": {
            "post": {
                "description": "Cancel an enrollment",
                "tags": [
                    "enrollments"
                ],
                "summary": "Cancel enrollment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enrollment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/enrollments/{id}/complete": {
            "post": {
                "description": "Mark an enrollment as completed",
                "tags": [
                    "enrollments"
                ],
                "summary": "Complete enrollment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enrollment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/enrollments/{id}/progress": {
            "put": {
                "description": "Update the progress percentage of an enrollment",
                "tags": [
                    "enrollments"
                ],
                "summary": "Update enrollment progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enrollment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Progress request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{user_id}/enrollments": {
            "get": {
                "description": "Retrieve paginated list of user enrollments with optional filters",
                "tags": [
                    "enrollments"
                ],
                "summary": "Get user enrollments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user enrollment",
                "tags": [
                    "enrollments"
                ],
                "summary": "Create user enrollment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enrollment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateEnrollmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CreateEnrollmentRequest": {
            "type": "object",
            "required": [
                "expires_at",
                "trade_id"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "trade_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateEnrollmentRequest": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "progress_pct": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}